+++ src/components/redux/const/actionTypes/Post.js
@@ -1,3 +1,4 @@
 export const POST_FETCH_REQUEST = 'POST_FETCH_REQUEST';
 export const POST_FETCH_SUCCESS = 'POST_FETCH_SUCCESS';
 export const POST_FETCH_ERROR = 'POST_FETCH_ERROR';
+export const POST_INC_LIKE = 'POST_INC_LIKE';
