+++ src/components/redux/reducers/Post.js
@@ -1,5 +1,5 @@
 import { POST_FETCH_REQUEST, POST_FETCH_SUCCESS,
-         POST_FETCH_ERROR } from '../const/actionTypes/Post';
+         POST_FETCH_ERROR, POST_INC_LIKE } from '../const/actionTypes/Post';
 
 const initialState = {
   isFetching: false,
@@ -12,9 +12,11 @@ export default function post(state = initialState, action) {
     case POST_FETCH_REQUEST:
       return Object.assign({}, initialState, {isFetching: true});
     case POST_FETCH_SUCCESS:
-      return Object.assign({}, initialState, {entry: action.response});
+      return Object.assign({}, initialState, {entry: action.response[0]});
     case POST_FETCH_ERROR:
       return Object.assign({}, initialState, {error: true});
+    case POST_INC_LIKE:
+      return Object.assign({}, initialState, {entry: action.item});
     default:
       return state;
   }
