+++ src/components/redux/store/index.js
@@ -1,6 +1,6 @@
 import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
-import reducers from 'components/reducers';
+import reducers from 'components/redux/reducers';
 import DevTools from 'containers/DevTools';
 
 const store = createStore(reducers,
