+++ src/containers/PostContainer.js
@@ -1,5 +1,6 @@
 import { connect } from 'react-redux';
 import Post from 'components/views/Post';
+import { incLike } from 'components/redux/actions/Post';
 
 const stateToProps = (state) => (
   {
@@ -9,4 +10,10 @@ const stateToProps = (state) => (
   }
 );
 
-export default connect(stateToProps)(Post);
+const actionToProps = (dispatch) => (
+  {
+    like: (id) => (dispatch(incLike(id)))
+  }
+);
+
+export default connect(stateToProps, actionToProps)(Post);
